major refactoring and auto_group initial commit
