+++ protector/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'Sergey Lihobabin'
-__version__ = '0.3.10'
+__version__ = '0.3.13'
 default_app_config = 'protector.apps.ProtectorConfig'
