+++ protector/models.py
@@ -1,70 +1,25 @@
 # -*- coding: utf-8 -*-
 from django.db import models
-from django.db.models.query import QuerySet
 from django.conf import settings
 from django.core.exceptions import PermissionDenied
 from django.contrib import auth
 from django.contrib.auth.models import Permission
-from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.fields import GenericRelation, GenericForeignKey
-from django.core.cache import cache
 from django.utils.translation import ugettext_lazy as _
 from mptt.models import MPTTModel, TreeForeignKey
+from protector.internals import NULL_OWNER_TO_PERMISSION_CTYPE_ID, \
+    NULL_OWNER_TO_PERMISSION_OBJECT_ID, DEFAULT_ROLE, \
+    ADD_PERMISSION_PERMISSION, VIEW_RESTRICTED_OBJECTS, \
+    VIEW_PERMISSION_NAME, get_user_ctype
+from protector.helpers import get_view_permission
+from protector.managers import GenericUserToGroupManager, OwnerToPermissionManager, \
+    OwnerPermissionManager, UserGroupManager, GroupUserManager, RestrictedManager
 
-
-def get_user_ctype():
-    return ContentType.objects.get_for_model(get_user_model())
-
-#  Need this to avoid null values in OwnerToPermission table
-NULL_OWNER_TO_PERMISSION_OBJECT_ID = 0
-NULL_OWNER_TO_PERMISSION_CTYPE_ID = 1  # That is ContentType ctype id
-
-ADD_PERMISSION_PERMISSION = 'add_permission'
-VIEW_RESTRICTED_OBJECTS = 'view_restricted_objects'
-
-VIEW_PERMISSION_NAME = 'protector.{0}'.format(VIEW_RESTRICTED_OBJECTS)
 
 #  Form a from clause for all permission related to their owners
 #  role of user in group must not be empty
 #  if permission roles is empty than it is applied to all roles in group
-
-DEFAULT_ROLE = 1
-
-
-def get_view_permission():
-    codename = VIEW_RESTRICTED_OBJECTS
-    if Restricted._view_perm is None:
-        ctype = ContentType.objects.get_for_model(OwnerToPermission)
-        Restricted._view_perm = Permission.objects.get(
-            codename=codename, content_type=ctype
-        )
-    return Restricted._view_perm
-
-
-def get_permission_owners_query():
-    """
-        This functions generate SQL statement for selecting groups and perms.
-        Sadly, Django doesn't support join in ORM
-        Should not select users that hasn't got any role
-        Should select perms that assigned to any role
-    """
-    owners_query = """
-        {group_table_name!s} gug LEFT JOIN
-        {owner_table_name!s} op ON
-            gug.group_id = op.owner_object_id AND
-            gug.group_content_type_id = op.owner_content_type_id AND
-            gug.roles & op.roles LEFT JOIN
-        {global_table_name!s} gl ON
-            gl.content_type_id = gug.group_content_type_id AND
-            gl.roles & gug.roles
-    """
-    return owners_query.format(
-        owner_table_name=OwnerToPermission._meta.db_table,
-        group_table_name=GenericUserToGroup._meta.db_table,
-        global_table_name=GenericGlobalPerm._meta.db_table,
-        null_owner_id=NULL_OWNER_TO_PERMISSION_OBJECT_ID
-    )
 
 
 class GenericGlobalPerm(models.Model):
@@ -83,18 +38,6 @@ class GenericGlobalPerm(models.Model):
         verbose_name = _('global group permission')
         verbose_name_plural = _('global group permissions')
         unique_together = ('content_type', 'permission')
-
-
-class GenericUserToGroupQuerySet(QuerySet):
-    def by_role(self, roles):
-        utg_table_name = self.model._meta.db_table
-        return self.extra(
-            where=["{utg_table}.roles & %s".format(utg_table=utg_table_name)],
-            params=[roles]
-        )
-
-
-GenericUserToGroupManager = models.Manager.from_queryset(GenericUserToGroupQuerySet)
 
 
 class GenericUserToGroup(models.Model):
@@ -131,17 +74,6 @@ class GenericUserToGroup(models.Model):
             group_id=self.group_id,
             username=self.user.username
         )
-
-
-class OwnerToPermissionQuerySet(QuerySet):
-    def without_obj_perms(self):
-        return self.filter(
-            object_id=NULL_OWNER_TO_PERMISSION_OBJECT_ID,
-            content_type_id=NULL_OWNER_TO_PERMISSION_CTYPE_ID
-        )
-
-
-OwnerToPermissionManager = models.Manager.from_queryset(OwnerToPermissionQuerySet)
 
 
 class OwnerToPermission(models.Model):
@@ -337,6 +269,14 @@ class AbstractGenericGroup(GenericPermsMixin):
     )
     DEFAULT_ROLE = PARTICIPANT
 
+    """
+        You could define a list of pairs "field_name, roles"
+        Field name should be a foreign key to user
+        to auto add this user to this group.
+        This is useful for cases like auto assign permissions to post author
+    """
+    MEMBER_FOREIGN_KEY_FIELDS = []
+
     users_relations = GenericRelation(
         GenericUserToGroup, content_type_field='group_content_type',
         object_id_field='group_id'
@@ -349,6 +289,11 @@ class AbstractGenericGroup(GenericPermsMixin):
         super(AbstractGenericGroup, self).__init__(*args, **kwargs)
         self.users = GroupUserManager(self)
 
+    def save(self, *args, **kwargs):
+        super(AbstractGenericGroup, self).save(*args, **kwargs)
+        for field, roles in self.MEMBER_FOREIGN_KEY_FIELDS:
+            self.users.add(getattr(self, field), roles=roles)
+
     def get_roles(self, user):
         try:
             user_roles = self.users_relations.values('roles').get(user=user)['roles']
@@ -356,114 +301,6 @@ class AbstractGenericGroup(GenericPermsMixin):
             return []
         else:
             return [role[0] for role in self.ROLES if role[0] & user_roles]
-
-
-def filter_queryset_by_permission(qset, user, permission):
-    perm_id = get_permission_id_by_name(permission)
-    if user.has_perm(permission):
-        return qset.all()
-    if user.id is None or perm_id is None:
-        return qset.none()
-    condition = _get_permission_filter(qset, user.id, perm_id)
-    return qset.extra(where=[condition])
-
-
-def _get_restriction_filter(qset, user_id, perm_id):
-    if hasattr(qset, 'get_restriction_id_field'):
-        obj_id_field = qset.get_restriction_id_field()
-    else:
-        obj_id_field = "{table_name!s}.restriction_id".format(table_name=qset.model._meta.db_table)
-    if hasattr(qset, 'get_restriction_ctype_id_field'):
-        ctype_id_field = qset.get_restriction_ctype_id_field()
-    else:
-        ctype_id_field = "{table_name!s}.restriction_content_type_id".format(
-            table_name=qset.model._meta.db_table
-        )
-    return _get_filter_by_perm_condition(qset, user_id, perm_id, obj_id_field, ctype_id_field)
-
-
-def _get_permission_filter(qset, user_id, perm_id):
-    if hasattr(qset, 'get_obj_id_field'):
-        obj_id_field = qset.get_obj_id_field()
-    else:
-        obj_id_field = "{table_name!s}.id".format(table_name=qset.model._meta.db_table)
-    if hasattr(qset, 'get_ctype_id_field'):
-        ctype_id_field = qset.get_ctype_id_field()
-    else:
-        ctype_id_field = str(ContentType.objects.get_for_model(qset.model).id)
-    return _get_filter_by_perm_condition(qset, user_id, perm_id, obj_id_field, ctype_id_field)
-
-
-def _generate_filter_condition(user_id, perm_id, ctype_id_field, obj_id_field):
-    condition = """
-        gug.user_id = {user_id!s} AND (
-            (
-                op.permission_id = {perm_id!s} AND
-                op.content_type_id = {ctype_id_field!s} AND
-                (op.object_id = {obj_id_field!s} OR op.object_id = {null_owner_id!s})
-            ) OR (
-                gl.permission_id = {perm_id!s} AND
-                gl.content_type_id = {ctype_id_field!s} AND
-                gug.group_id = {obj_id_field!s}
-            )
-        )
-    """
-    return condition.format(
-        user_id=user_id, perm_id=perm_id,
-        null_owner_id=NULL_OWNER_TO_PERMISSION_OBJECT_ID,
-        obj_id_field=obj_id_field,
-        ctype_id_field=ctype_id_field
-    )
-
-
-def _get_filter_by_perm_condition(qset, user_id, perm_id, obj_id_field, ctype_id_field):
-    # here we brake some rules about sql sanitizing
-    # it is a shame, but this is an internal function so we can live with it
-    condition = "EXISTS (SELECT op.id FROM {permission_owners} WHERE {filter_condition})"
-    return condition.format(
-        permission_owners=get_permission_owners_query(),
-        filter_condition=_generate_filter_condition(
-            user_id, perm_id, ctype_id_field, obj_id_field
-        )
-    )
-
-
-class PermissionQuerySet(QuerySet):
-    """
-        Queryset is used for filtering any queryset by user permissions on its objects
-    """
-    def filter_by_permission(self, user, permission):
-        return filter_queryset_by_permission(self, user, permission)
-
-
-PermissionedManager = models.Manager.from_queryset(PermissionQuerySet)
-
-
-class RestrictedQuerySet(PermissionQuerySet):
-    """
-        Queryset is used for filtering objects not visible to user
-    """
-
-    def visible(self, user=None):
-        if user is None:
-            return self.filter(restriction_id__isnull=True)
-        if user.has_perm(VIEW_PERMISSION_NAME):
-            return self
-        if user.id is None:
-            return self.filter(restriction_id__isnull=True)
-        condition = self._get_visible_condition(user.id, get_view_permission().id)
-        return self.extra(where=[condition])
-
-    def _get_visible_condition(self, user_id, perm_id):
-        condition = """
-            {table_name!s}.restriction_id IS NULL OR
-        """
-        condition += _get_restriction_filter(self, user_id, perm_id)
-        condition = condition.format(table_name=self.model._meta.db_table)
-        return condition
-
-
-RestrictedManager = models.Manager.from_queryset(RestrictedQuerySet)
 
 
 class Restriction(MPTTModel, models.Model):
@@ -510,7 +347,6 @@ class Restricted(models.Model):
     restriction = GenericForeignKey('restriction_content_type', 'restriction_id')
 
     objects = RestrictedManager()
-    _view_perm = None
 
     class Meta:
         abstract = True
@@ -629,207 +465,3 @@ class Restricted(models.Model):
         if not created and otp.roles != roles:
             otp.roles |= roles
             otp.save()
-
-
-def get_default_group_ctype():
-    return ContentType.objects.get_by_natural_key(
-        *settings.PROTECTOR_GENERIC_GROUP.lower().split('.')
-    )
-
-
-class UserGroupManager(models.Manager):
-
-    def __init__(self, instance):
-        super(UserGroupManager, self).__init__()
-        self.instance = instance
-
-    def add(self, group, responsible=None, roles=None):
-        roles = roles or group.DEFAULT_ROLE
-        utg, created = GenericUserToGroup.objects.get_or_create(
-            user=self.instance,
-            group_id=group.pk,
-            group_content_type=ContentType.objects.get_for_model(group),
-            defaults={'responsible': responsible, 'roles': roles}
-        )
-        if not created and utg.roles != roles:
-            utg.roles |= roles
-            utg.save()
-
-    def remove(self, group, roles=None):
-        try:
-            utg = GenericUserToGroup.objects.get(
-                group_id=group.pk,
-                group_content_type=ContentType.objects.get_for_model(group),
-                user=self.instance
-            )
-        except GenericUserToGroup.DoesNotExist:
-            return
-        if roles is None or utg.roles == roles:
-            utg.delete()
-        else:
-            utg.roles &= ~roles
-            utg.save()
-
-    def get_queryset(self, group_type=None):
-        if group_type is None:
-            group_type = get_default_group_ctype()
-        return group_type.model_class().objects.filter(
-            pk__in=GenericUserToGroup.objects.filter(
-                group_content_type=group_type,
-                user=self.instance
-            ).values_list('group_id')
-        )
-
-    def by_ctype(self, group_type, roles=None):
-        group_model = group_type.model_class()
-        utg_qset = GenericUserToGroup.objects.filter(
-            group_content_type=group_type,
-            user=self.instance
-        )
-        if roles is not None:
-            utg_qset = utg_qset.extra(where=["roles & %s"], params=[roles])
-        return group_model.objects.filter(
-            pk__in=utg_qset.values_list('group_id')
-        )
-
-
-class GroupUserManager(models.Manager):
-
-    def __init__(self, instance):
-        super(GroupUserManager, self).__init__()
-        self.instance = instance
-
-    def get_queryset(self):
-        user_ids = self.instance.users_relations.values_list('user_id', flat=True)
-        return get_user_model().objects.filter(id__in=user_ids)
-
-    def add(self, user, roles=None, responsible=None):
-        roles = roles or self.instance.DEFAULT_ROLE
-        gug, created = GenericUserToGroup.objects.get_or_create(
-            user=user, group_id=self.instance.id,
-            group_content_type=ContentType.objects.get_for_model(self.instance),
-            defaults={'roles': roles, 'responsible': responsible}
-        )
-        if not created:
-            gug.roles |= roles
-            gug.save()
-
-    def remove(self, user, roles=None):
-        # if roles is None just remove user from group else remove role from user
-        try:
-            utg = GenericUserToGroup.objects.get(
-                group_id=self.instance.pk,
-                group_content_type=ContentType.objects.get_for_model(self.instance),
-                user=user
-            )
-        except GenericUserToGroup.DoesNotExist:
-            return
-        if roles is None or utg.roles == roles:
-            utg.delete()
-        else:
-            utg.roles &= ~roles
-            utg.save()
-
-    def by_role(self, roles):
-        links = self.instance.users_relations.all()
-        if roles is None:
-            links = links.filter(roles__isnull=True)
-        else:
-            links = links.extra(where=["roles & %s"], params=[roles])
-        return get_user_model().objects.filter(id__in=links.values_list('user_id', flat=True))
-
-
-class OwnerPermissionManager(models.Manager):
-
-    def __init__(self, instance):
-        super(OwnerPermissionManager, self).__init__()
-        self.instance = instance
-
-    def get_queryset(self):
-        ctype = ContentType.objects.get_for_model(self.instance)
-        return Permission.objects.filter(
-            generic_restriction_relations__owner_object_id__in=[self.instance.pk],
-            generic_restriction_relations__owner_content_type__in=[ctype]
-        ).distinct()
-
-    def add(self, perm, obj=None, responsible=None, roles=None):
-        roles = roles or DEFAULT_ROLE
-        kwargs = {
-            'owner_object_id': self.instance.id,
-            'owner_content_type': ContentType.objects.get_for_model(self.instance),
-            'defaults': {'responsible': responsible, 'roles': roles}
-        }
-        if isinstance(perm, str):
-            kwargs['permission_id'] = get_permission_id_by_name(perm)
-        else:
-            kwargs['permission'] = perm
-
-        if obj is not None:
-            kwargs['object_id'] = obj.pk
-            kwargs['content_type'] = ContentType.objects.get_for_model(obj)
-        else:
-            kwargs['object_id'] = NULL_OWNER_TO_PERMISSION_OBJECT_ID
-            kwargs['content_type'] = ContentType.objects.get_for_id(
-                NULL_OWNER_TO_PERMISSION_CTYPE_ID
-            )
-        otp, created = OwnerToPermission.objects.get_or_create(
-            **kwargs
-        )
-        if not created and otp.roles != roles:
-            otp.roles |= roles
-            otp.save()
-
-    def remove(self, perm, obj=None, roles=None):
-        if obj is None:
-            obj_id = NULL_OWNER_TO_PERMISSION_OBJECT_ID
-            obj_ctype_id = NULL_OWNER_TO_PERMISSION_CTYPE_ID
-        else:
-            obj_id = obj.pk
-            obj_ctype_id = ContentType.objects.get_for_model(obj)
-
-        if isinstance(perm, str):
-            perm_id = get_permission_id_by_name(perm)
-        else:
-            perm_id = perm.id
-
-        try:
-            otp = OwnerToPermission.objects.get(
-                permission_id=perm_id, owner_object_id=self.instance.pk,
-                owner_content_type=ContentType.objects.get_for_model(self.instance),
-                object_id=obj_id, content_type_id=obj_ctype_id
-            )
-        except OwnerToPermission.DoesNotExist:
-            return
-        if roles is None or otp.roles == roles:
-            otp.delete()
-        else:
-            otp.roles &= ~roles
-            otp.save()
-
-
-def get_permission_id_by_name(permission):
-    cache_key = 'permission_id_cache_' + permission
-    perm_id = cache.get(cache_key, None)
-    if perm_id is None:
-        try:
-            perm_id = Permission.objects.get(
-                codename=permission.split('.')[1],
-                content_type__app_label=permission.split('.')[0]
-            ).id
-        except Permission.DoesNotExist:
-            return None
-        cache.set(cache_key, perm_id)
-    return perm_id
-
-
-class PermAnnotatedMixin(QuerySet):
-
-    annotate_perms_for_user = None
-
-    def annotate_perms(self, user):
-        new_qset = self.all()
-        new_qset.annotate_perms_for_user = user
-        return new_qset
-
-    def _fetch_all(self):
-        super(PermAnnotatedMixin, self)._fetch_all()
