+++ runtests.py
@@ -12,5 +12,5 @@ import django
 def runtests():
     django.setup()
     test_runner = get_runner(settings)(verbosity=1, interactive=True)
-    failures = test_runner.run_tests(['protector'])
+    failures = test_runner.run_tests(['protector', 'test_app'])
     sys.exit(bool(failures))
