+++ test_project/test_app/models.py
@@ -1,21 +1,21 @@
 from django.db import models
-from django.conf import settings
-from protector.models import AbstractGenericGroup, Restricted, UserGenericPermsMixin,\
-    PermissionedManager
+from protector.models import AbstractGenericGroup, Restricted, UserGenericPermsMixin
+from protector.managers import PermissionedManager
 from django.contrib.auth.models import AbstractBaseUser, UserManager
 
 
 class TestGroup(AbstractGenericGroup, Restricted):
     name = models.CharField(max_length=100)
-    
+
     by_perm = PermissionedManager()
-    
+
     class Meta:
         verbose_name = u'Test Group'
 
+
 class TestUser(UserGenericPermsMixin, AbstractBaseUser):
     USERNAME_FIELD = 'username'
-    REQUIRED_FIELDS = ('email', )  
+    REQUIRED_FIELDS = ('email', )
     username = models.CharField(max_length=30, unique=True)
     is_staff = models.BooleanField(
         default=False,
@@ -26,9 +26,28 @@ class TestUser(UserGenericPermsMixin, AbstractBaseUser):
     email = models.CharField(max_length=50)
 
     objects = UserManager()
-    
+
     def get_short_name(self):
         return self.username
 
     def get_full_name(self):
         return self.username
+
+
+class TestPost(AbstractGenericGroup):
+    SUBSCRIBER = 1
+    AUTHOR = 2
+    ROLES = (
+        (SUBSCRIBER, 'Subscriber'),
+        (AUTHOR, 'Author')
+    )
+    author = models.ForeignKey(to=TestUser)
+
+    MEMBER_FOREIGN_KEY_FIELDS = (
+        ('author', AUTHOR),
+    )
+
+    class Meta:
+        permissions = (
+            ('manage_post', 'Manage Post'),
+        )
